scripts/requirements.txt
asyncpg>=0.29.0
redis>=5.0.0
alpaca-py>=0.42.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
pandas-market-calendars>=4.3.0
pytz>=2023.3
requests>=2.31.0

